1. 简短重构，将使用 Mono.Cecil 生成程序集的逻辑单独分离到 CecilAssemblyBuilder 中，方便后续重构。 2. 解决程序集卸载问题。
